c323: extract Routed-classifier family from api/client.ts to api/routed.ts (#441 P1 2.4)
api/client.ts 1133 LOC -> 1095 LOC (-38 LOC). New file: api/routed.ts (70 LOC).
Moved RoutedFoldMetric + RoutedMethodMetrics + TopicRoutedClassifier +
TopicRoutedDeepGate types and the 2 runtime methods. client.ts re-exports
types for back-compat with existing consumers; api.topicRoutedClassifier /
api.topicRoutedDeepGate delegate to routed.* under the hood.
typecheck clean, 44/44 tests pass.
